+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/21 11:48:53 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/02/14 13:49:21 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/02/21 13:34:45 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 ** Returns a pointer to 'dst' memory area or 'NULL' if src is empty.
 */
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char			*d;
 	const char		*s;
